[GEOT-7344] Child feature ClientProperties are duplicated on no-GML XML element parent container from the child elements
