+++ modules/extension/app-schema/app-schema/src/test/resources/test-data/stations-nestednogml/measurementsDefaultGeometry.xml
@@ -1,0 +1,68 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<as:AppSchemaDataAccess xmlns:as="http://www.geotools.org/app-schema"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://www.geotools.org/app-schema AppSchemaDataAccess.xsd">
+    <namespaces>
+        <Namespace>
+            <prefix>ms</prefix>
+            <uri>http://www.measurements.org/1.0</uri>
+        </Namespace>
+    </namespaces>
+    <sourceDataStores>
+        <DataStore>
+            <id>measurementsDefaultGeometry</id>
+            <parameters>
+                <Parameter>
+                    <name>directory</name>
+                    <value>file:./</value>
+                </Parameter>
+            </parameters>
+        </DataStore>
+    </sourceDataStores>
+    <targetTypes>
+        <FeatureType>
+            <schemaUri>measurementsDefaultGeometry.xsd</schemaUri>
+        </FeatureType>
+    </targetTypes>
+    <typeMappings>
+        <FeatureTypeMapping>
+            <mappingName>measurementsCode</mappingName>
+            <sourceDataStore>measurementsDefaultGeometry</sourceDataStore>
+            <sourceType>measurementsDefaultGeometry</sourceType>
+            <targetElement>ms:MeasurementCode</targetElement>
+            <attributeMappings>
+                <AttributeMapping>
+                    <targetAttribute>ms:MeasurementCode</targetAttribute>
+                    <idExpression>
+                        <OCQL>ID</OCQL>
+                    </idExpression>
+                    <ClientProperty>
+                        <name>codename</name>
+                        <value>strConcat(NAME, '-code')</value>
+                    </ClientProperty>
+                </AttributeMapping>
+                <AttributeMapping>
+                    <targetAttribute>ms:name</targetAttribute>
+                    <sourceExpression>
+                        <OCQL>NAME</OCQL>
+                    </sourceExpression>
+                    <ClientProperty>
+                        <name>code</name>
+                        <value>strConcat(NAME, '-code')</value>
+                    </ClientProperty>
+                </AttributeMapping>
+                <AttributeMapping>
+                    <targetAttribute>ms:unit</targetAttribute>
+                    <sourceExpression>
+                        <OCQL>UNIT</OCQL>
+                    </sourceExpression>
+                </AttributeMapping>
+                <AttributeMapping>
+                    <targetAttribute>FEATURE_LINK[1]</targetAttribute>
+                    <sourceExpression>
+                        <OCQL>STATION</OCQL>
+                    </sourceExpression>
+                </AttributeMapping>
+            </attributeMappings>
+        </FeatureTypeMapping>
+    </typeMappings>
+</as:AppSchemaDataAccess>
