+++ modules/extension/app-schema/app-schema/src/test/resources/test-data/stations-nestednogml/measurementsManyToOne.properties
@@ -1,0 +1,4 @@
+_=ID:String,NAME:String,UNIT:String,STATION:String,SAMPLEDAREA_CODE:String,SAMPLEDAREA_GEOM:Geometry:srid=4326
+ms.1=ms.1|temperature|c|st.1|SA1234|POLYGON((-2 2, 0 2, 0 -2, -2 -2, -2 2))
+ms.2=ms.2|wind|km/h|st.1|SA1234|POLYGON((-4 1, -2 1, -2 -3, -4 -3, -4 1))
+ms.3=ms.3|pressure|kPa|st.1|SA5678|POLYGON((-3 3, -1 3, -1 -1, -3 -1, -3 3))
