+++ modules/extension/app-schema/app-schema/src/test/resources/test-data/stations-nestednogml/stationsDefaultGeometry.properties
@@ -1,0 +1,4 @@
+_=ID:String,NAME:String,LOCATION:String,ST_POSITION:Geometry:srid=4326
+st.1=st.1|station1|Europe|POINT(-1 1)
+st.2=st.2|station2|Asia|POINT(-2 1)
+st.3=st.3|station3|Australia|POINT(-3 1)
