+++ modules/extension/app-schema/app-schema/src/test/resources/test-data/stations-nestednogml/stationsDefaultGeometry.xml
@@ -1,0 +1,72 @@
+<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+<as:AppSchemaDataAccess xmlns:as="http://www.geotools.org/app-schema">
+    <namespaces>
+        <Namespace>
+            <prefix>st</prefix>
+            <uri>http://www.stations.org/1.0</uri>
+        </Namespace>
+        <Namespace>
+            <prefix>ms</prefix>
+            <uri>http://www.measurements.org/1.0</uri>
+        </Namespace>
+    </namespaces>
+    <sourceDataStores>
+        <DataStore>
+            <id>stationsDefaultGeometry</id>
+            <parameters>
+                <Parameter>
+                    <name>directory</name>
+                    <value>file:./</value>
+                </Parameter>
+            </parameters>
+        </DataStore>
+    </sourceDataStores>
+    <targetTypes>
+        <FeatureType>
+            <schemaUri>stationsDefaultGeometry.xsd</schemaUri>
+        </FeatureType>
+    </targetTypes>
+    <typeMappings>
+        <FeatureTypeMapping>
+            <mappingName>stationsCode</mappingName>
+            <sourceDataStore>stationsDefaultGeometry</sourceDataStore>
+            <sourceType>stationsDefaultGeometry</sourceType>
+            <targetElement>st:StationWithMeasurementCode</targetElement>
+            <defaultGeometry>st:location/st:position</defaultGeometry>
+            <attributeMappings>
+                <AttributeMapping>
+                    <targetAttribute>st:StationWithMeasurementCode</targetAttribute>
+                    <idExpression>
+                        <OCQL>ID</OCQL>
+                    </idExpression>
+                </AttributeMapping>
+                <AttributeMapping>
+                    <targetAttribute>st:name</targetAttribute>
+                    <sourceExpression>
+                        <OCQL>NAME</OCQL>
+                    </sourceExpression>
+                </AttributeMapping>
+                <AttributeMapping>
+                    <targetAttribute>st:location/st:name</targetAttribute>
+                    <sourceExpression>
+                        <OCQL>LOCATION</OCQL>
+                    </sourceExpression>
+                </AttributeMapping>
+                <AttributeMapping>
+                    <targetAttribute>st:location/st:position</targetAttribute>
+                    <sourceExpression>
+                        <OCQL>ST_POSITION</OCQL>
+                    </sourceExpression>
+                </AttributeMapping>
+                <AttributeMapping>
+                    <targetAttribute>st:measurements</targetAttribute>
+                    <sourceExpression>
+                        <OCQL>ID</OCQL>
+                        <linkElement>measurementsCode</linkElement>
+                        <linkField>FEATURE_LINK[1]</linkField>
+                    </sourceExpression>
+                </AttributeMapping>
+            </attributeMappings>
+        </FeatureTypeMapping>
+    </typeMappings>
+</as:AppSchemaDataAccess>
