+++ modules/extension/app-schema/app-schema/src/test/resources/test-data/stations-nestednogml/stationsDefaultGeometryNonExistentProperty.xml
@@ -1,0 +1,59 @@
+<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+<as:AppSchemaDataAccess xmlns:as="http://www.geotools.org/app-schema">
+    <namespaces>
+        <Namespace>
+            <prefix>st</prefix>
+            <uri>http://www.stations.org/1.0</uri>
+        </Namespace>
+    </namespaces>
+    <sourceDataStores>
+        <DataStore>
+            <id>stationsDefaultGeometry</id>
+            <parameters>
+                <Parameter>
+                    <name>directory</name>
+                    <value>file:./</value>
+                </Parameter>
+            </parameters>
+        </DataStore>
+    </sourceDataStores>
+    <targetTypes>
+        <FeatureType>
+            <schemaUri>stationsDefaultGeometry.xsd</schemaUri>
+        </FeatureType>
+    </targetTypes>
+    <typeMappings>
+        <FeatureTypeMapping>
+            <sourceDataStore>stationsDefaultGeometry</sourceDataStore>
+            <sourceType>stationsDefaultGeometry</sourceType>
+            <targetElement>st:Station</targetElement>
+            <defaultGeometry>st:location/st:notThere</defaultGeometry>
+            <attributeMappings>
+                <AttributeMapping>
+                    <targetAttribute>st:Station</targetAttribute>
+                    <idExpression>
+                        <OCQL>ID</OCQL>
+                    </idExpression>
+                </AttributeMapping>
+                <AttributeMapping>
+                    <targetAttribute>st:name</targetAttribute>
+                    <sourceExpression>
+                        <OCQL>NAME</OCQL>
+                    </sourceExpression>
+                </AttributeMapping>
+                <AttributeMapping>
+                    <targetAttribute>st:location/st:name</targetAttribute>
+                    <sourceExpression>
+                        <OCQL>LOCATION</OCQL>
+                    </sourceExpression>
+                </AttributeMapping>
+                <AttributeMapping>
+                    <targetAttribute>st:location/st:position</targetAttribute>
+                    <sourceExpression>
+                        <OCQL>ST_POSITION</OCQL>
+                    </sourceExpression>
+                </AttributeMapping>
+            </attributeMappings>
+        </FeatureTypeMapping>
+    </typeMappings>
+</as:AppSchemaDataAccess>
